requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.4.12
albumentations>=1.0.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0
Pillow>=8.0.0
Flask>=2.0.0
numpy>=1.21.0